rename Binary to ByteStream
