+++ src/Lib/ByteStream/ByteReaderInterface.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace PhpProfiler\Lib\Binary;
+namespace PhpProfiler\Lib\ByteStream;
 
 use ArrayAccess;
 use LogicException;
